pyproject.toml
[project]
name = "metaimpro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.1",
    "mido>=1.3.3",
    "numpy>=2.2.4",
    "polars>=1.27.1",
    "pyfluidsynth>=1.3.4",
    "pygame>=2.6.1",
    "python-rtmidi>=1.5.8",
    "tqdm>=4.67.1",
]
